+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Re-enable API write access.
 - Use newly migrated DB (using officeId).
 - Require officeId to be specified in config (to avoid renaming issues).
+- Disable public Cognito registrations.
 
 ## [1.2.0] - 2020-09-14